refactor(daemon): extract daemon main into zccache::daemon::entry (#997)
Move the standalone daemon's main body (Args, tracing setup, pre-bind
probe, bind-race handling, background loaders, run_server) out of
src/bin/zccache-daemon.rs into a new library module daemon::entry,
gated behind a new daemon-entry feature (clap + tracing-subscriber).
The bin is now a 3-line shim that installs the global allocator (which
must live in the final binary) and calls entry::run().

daemon-bin now pulls daemon-entry; cli pulls it too so the zccache
binary can route to the daemon via argv[0] dispatch (#998). run() is
split into run()/run_with(Args) so dispatch and tests can drive the
daemon without re-parsing process argv. Zero behavior change.

The arg parser is now library-testable (2 unit tests added) — it lived
in a bin and could not be unit-tested before.

Prereq for #998 (argv[0] dispatch) and #999 (self-materialization).
Closes #997
